add run tests by setup file

+++ tree/tree.py
@@ -1,10 +1,10 @@
 import os
 
 from tree.exceptions import (
-    TreeConfigNoneExist,
     TreeFileNoneExist,
+    TreeDirNoneExist,
 )
-from .utils import (
+from tree.utils import (
     hash_to_path,
     hashing_file,
 )
@@ -15,14 +15,14 @@ class TreeStorage:
     Main object for the manipulate of the tree storage
     """
 
-    def __init__(self, config=None):
+    def __init__(self, path=None):
         """
         When initialize TreeStorage client
         must be initialize TreeConfig
         """
-        if not config:
-            raise TreeConfigNoneExist
-        self.config = config
+        self.path_to_tree = path
+        if not self.path_to_tree:
+            raise TreeDirNoneExist
         self.file_hash_name = ''
 
     def breed(self, file_byte, mode='wb'):
@@ -32,8 +32,8 @@ class TreeStorage:
         try:
             self.file_hash_name = hashing_file(file_byte)
             hash_dir, hash_file = hash_to_path(hash_name=self.file_hash_name)
-            os.makedirs(self.config.PATH_TO_TREE + hash_dir)
-            with open(os.path.join(self.config.PATH_TO_TREE, hash_dir, hash_file),
+            os.makedirs(os.path.join(self.path_to_tree, hash_dir))
+            with open(os.path.join(self.path_to_tree, hash_dir, hash_file),
                       mode=mode) as file_hash:
                 return file_hash.write(file_byte)
         except IOError as error:
@@ -43,11 +43,9 @@ class TreeStorage:
         """
         Remove file from the tree storage
         """
-        try:
-            if greedy:
-                pass  # TODO: if dir is empty must be remove
-            hash_dir, hash_file = hash_to_path(hash_name=file_hash_name)
-            path = os.path.join(self.config.PATH_TO_TREE, hash_dir, hash_file)
+        if greedy:
+            pass  # TODO: if dir is empty must be remove
+        hash_dir, hash_file = hash_to_path(hash_name=file_hash_name)
+        path = os.path.join(self.path_to_tree, hash_dir, hash_file)
+        if os.path.isfile(path):
             return os.remove(path)
-        except TreeFileNoneExist(file_hash_name=file_hash_name) as error:
-            return error.msg
